Add solution for exercise 4_21

+++ part4/bloglist_backend/controllers/blogs.js
@@ -4,52 +4,60 @@ const User = require('../models/user')
 const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog
-        .find({}).populate('user', { username: 1, name: 1 })
+  const blogs = await Blog
+    .find({}).populate('user', { username: 1, name: 1 })
 
-    response.json(blogs)
+  response.json(blogs)
 })
 
 blogsRouter.post('/', async (request, response) => {
-    const body = request.body
-    const decodedToken = jwt.verify(request.token, process.env.SECRET)
-    if (!request.token || !decodedToken.id) {
-        return response.status(401).json({ error: 'token missing or invalid' })
-    }
-    const user = await User.findById(decodedToken.id)
+  const body = request.body
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
 
-    const blog = new Blog({
-        title: body.title,
-        author: body.author, 
-        url: body.url,
-        likes: body.likes,
-        user: user._id
-    })
-    const savedBlog = await blog.save()
+  const blog = new Blog({
+    title: body.title,
+    author: body.author, 
+    url: body.url,
+    likes: body.likes,
+    user: user._id
+  })
+  const savedBlog = await blog.save()
     
-    if (blog.url === undefined || blog.title === undefined) {
-        response.status(400).json({ error: 'url or title missing'})
-    } else {
-        user.blogs = user.blogs.concat(savedBlog._id)
-        await user.save()
-        response.status(201).json(savedBlog)
-    }
+  if (blog.url === undefined || blog.title === undefined) {
+    response.status(400).json({ error: 'url or title missing'})
+  } else {
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+    response.status(201).json(savedBlog)
+  }
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+  const blog = await Blog.findById(request.params.id)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  //console.log(blog.user)
+
+  if (blog.user.toString() !== decodedToken.id.toString()) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  } 
+
+  await Blog.findByIdAndRemove(request.params.id)
+  response.status(204).end()
 })
 
 blogsRouter.put('/:id', async (request, response) => {
-    const body = request.body
+  const body = request.body
 
-    const blog = {
-        likes: body.likes,
-    }
+  const blog = {
+    likes: body.likes,
+  }
 
-    const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
-    response.json(updatedBlog)
+  const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
+  response.json(updatedBlog)
 })
 
 module.exports = blogsRouter
